requirements.txt
pydantic>=2.0.0
pytest==8.2.2
google-adk
google-genai
